requirements.txt
asgiref==3.7.2
dj-rest-auth==4.0.1
Django==4.2.4
django-filter==23.2
django-reversion==5.0.4
djangorestframework==3.14.0
Markdown==3.4.4
pytz==2023.3
sqlparse==0.4.4
